Add support for "%s" in bundle name (replaced by hash)
This patch looks for '%s' in a named bundle and replaces it by the
bundles' hash.

This allows to have named bundles, whose path/url changes if the its
content changes.

You might want to refactor this into a getNameForType method (handling
both null values for $name and %s in a regular name).